requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
autopep8==2.0.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.4
django-author==1.2.0
django-celery-beat==2.3.0
django-cors-headers==3.11.0
django-debug-toolbar==3.4.0
django-import-export==2.9.0
django-import-export-celery==1.3
django-jazzmin==2.6.0
django-rest-passwordreset==1.2.1
django-simple-history==3.0.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
flower==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
humanize==4.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
setuptools-git==1.2
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tablib==3.2.1
tomli==2.0.1
tornado==6.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
watchdog==2.1.8
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
